requirements.txt
alembic==1.10.4
bcrypt==4.0.1
beautifulsoup4==4.12.2
cryptography==41.0.6
fastapi==0.95.2
Jinja2==3.1.2
passlib==1.7.4
psycopg2-binary==2.9.6
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
SQLAlchemy==2.0.12
tldextract==3.4.4
uvicorn==0.22.0